backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["maskeen  <120341809+mxskeen@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
feedparser = "^6.0.11"
supabase = "^2.16.0"
python-dotenv = "^1.1.1"
fastapi = "^0.115.14"
uvicorn = "^0.35.0"
bs4 = "^0.0.2"
requests = "^2.32.4"
sentence-transformers = "^4.1.0"
torch = "^2.7.1"
torchvision = "^0.22.1"
torchaudio = "^2.7.1"
openai = "^1.93.0"
trafilatura = "^2.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
